Fix formatting and naming on pie chart. rename file
