+++ src/components/dashboard/DashboardChart.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import { Doughnut } from "react-chartjs-2";
-import {} from "chart.js";
-import "chart.js/auto";
+import { ChartData } from "chart.js";
 import { Theme, useTheme } from "@mui/material/styles";
 import { Box, Typography } from "@mui/material";
 
@@ -11,15 +10,14 @@ import { Box, Typography } from "@mui/material";
  * @returns returns a pie chart that shows the pickup status of the orders
  */
 export default function DashboardChart() {
-    const myTheme = useTheme();
+    const theme = useTheme();
 
     /**
      * 
      * @param theme theme is used to set the background color of the doughnut chart
      * @returns returns the data for the doughnut chart
      */
-    function generateData(them: Theme): import("chart.js").ChartData<"doughnut", number[], unknown> {
-
+    function generateData(theme: Theme): ChartData<"doughnut", number[], unknown> {
         return {
             labels: [
                 "Picked Up",
@@ -29,8 +27,8 @@ export default function DashboardChart() {
                 label: "Status",
                 data: [90, 69],
                 backgroundColor: [
-                    them.palette.primary.main,
-                    them.palette.secondary.main
+                    theme.palette.primary.main,
+                    theme.palette.secondary.main
                 ],
                 hoverOffset: 4
             }],
@@ -39,10 +37,19 @@ export default function DashboardChart() {
 
     return (
         <React.Fragment>
-            <Typography sx={{fontWeight:"bold",marginBottom:"2%",paddingLeft:"0.6250em", textAlign: "left", fontSize:"1.0em"}}>Big Pick Up Status <a href="/orders">orders</a></Typography>
-            <Box sx={{height:"21.5vh", marginTop:"0", paddingBottom:"0.6250em"}}>
+            <Typography
+                sx={{
+                    fontWeight: "bold",
+                    marginBottom: "2%",
+                    paddingLeft: "0.6250em",
+                    textAlign: "left",
+                    fontSize: "1.0em"
+                }} >
+                Big Pick Up Status <a href="/orders">orders</a>
+            </Typography>
+            <Box sx={{ height: "21.5vh", marginTop: "0", paddingBottom: "0.6250em" }}>
                 <Doughnut
-                    data={generateData(myTheme)}
+                    data={generateData(theme)}
                     options={{
                         responsive: true,
                         maintainAspectRatio: false,
