+++ src/components/dashboard/customAdminPage.tsx
@@ -3,7 +3,7 @@ import { experimentalStyled as styled } from "@mui/material/styles";
 import Box from "@mui/material/Box";
 import Paper from "@mui/material/Paper";
 import Grid from "@mui/material/Grid";
-import DashboardChart from "./CustomPieChart";
+import DashboardChart from "./DashboardChart";
 import Typography from "@mui/material/Typography";
 import NewStrong from "./customStrongTag";
 "";
